requirement.txt
flask
python-dotenv
groq
weaviate-client
PyPDF2
python-docx
werkzeug
requests